pyproject.toml
[tool.poetry]
name = "faster"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.8"
FastAPI-SQLAlchemy = "^0.2.1"
alembic = "^1.5.4"
uvicorn = "^0.13.3"
fastapi = "^0.63.0"
SQLAlchemy = "^1.3.23"
psycopg2 = "^2.8.6"
"discord.py" = "^1.6.0"
sqlalchemy-get-or-create = "^0.1.5"
fastapi-pagination = "^0.5.2"
pydantic = "^1.7.3"
colorhash = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
